[Justice Counts] Make it possible to update an existing AgencySetting (Recidiviz/recidiviz-data#17150)
## Description of the change

Currently, we're getting the following error if you try to update a
existing Agency Setting:

```
Traceback (most recent call last):
  File "/app/.venv/lib/python3.9/site-packages/sqlalchemy/engine/base.py", line 1900, in _execute_context
    self.dialect.do_execute(
  File "/app/.venv/lib/python3.9/site-packages/sqlalchemy/engine/default.py", line 736, in do_execute
    cursor.execute(statement, parameters)
psycopg2.errors.UniqueViolation: duplicate key value violates unique constraint "unique_agency_setting"
```

This is because we [always try to create a new
AgencySetting](https://github.com/Recidiviz/recidiviz-data/blob/6ac3f9a77077bed2ba5363cbffec6441ef8be774/recidiviz/justice_counts/control_panel/routes/api.py#L105)
when the endpoint is hit. If there is an existing one with the same
source ID and type, Postgres throws an exception, because we're not
allowed to have to AgencySettings with the same source ID and type.

The fix is to update, rather than insert, when an AgencySetting of the
same source ID and type already exists in the DB. We can do that via
"insert on conflict update" (see
[here](https://docs.sqlalchemy.org/en/20/dialects/postgresql.html#insert-on-conflict-upsert)
for details).
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5cbd0d7da457221f988d93acc9beff0a6cc20d77