drop US_PA rows from StateSupervisionViolatedConditionEntry table and rebuild from StateSupervisionViolatedTypeEntry (Recidiviz/recidiviz-data#17075)
## Description of the change

The condition raw text field is supposed to be the violation code, but
in PA the violation codes have been mapped to integers 1-7, preventing
us from running a migration where we populate condition with the correct
enum based off of what's in condition raw text. This PR will drop the
US_PA rows from StateSupervisionViolatedConditionEntry, and then rebuild
them based off of US_PA entries in StateSupervisionViolatedTypeEntry.
This is an intermediate migration that will enable
https://github.com/Recidiviz/recidiviz-data/issues/16539

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 236a78cb3caab07b91091a6db0020368d3729165