[US_ME] Update ME SCCP resident record (Recidiviz/recidiviz-data#17153)
## Description of the change

- We updated the ME SCCP resident record so that admission and release
dates come from `CIS_319`, the table we use to create the criteria
queries that contain the most up-to-date release dates.
- Modified `served_x_portion_of_sentence.py `and
`x_monts_remaining_on_sentence` so that if the longest sentence ends in
NULL, it is prioritized and taken as the final release date (this will
prevent us from surfacing individuals to SCCP who don't have a release
date)
- Modified raw_table_import so that `CIS_319` is imported without magic
end/start dates and `raw_data_up_to_date_views` is changed to
`us_me_raw_data_up_to_date` to be consistent with the rest of our
codebase (this last piece was also refactored in the criteria queries)

These changes drop the number of eligible people from 249 to 222.

## Related issues

Closes Recidiviz/recidiviz-data#2775

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7834080b1e1dfdf2a9707eac424df734c1bc8f9d