[Security] Enabling JSON Parsing and Verbose logging on WAF (Recidiviz/recidiviz-data#17122)
## Description of the change

> The WAF in staging is not in enforcement mode. But if it was in
enforcement mode, it would have tried to block this
[request](https://console.cloud.google.com/logs/query;cursorTimestamp=2022-12-06T23:49:22.026643Z;query=resource.type:%2528http_load_balancer%2529%20AND%20jsonPayload.enforcedSecurityPolicy.name:%2528recidiviz-staging-waf%2529%0AhttpRequest.requestUrl%3D%22https:%2F%2Fapp-staging.recidiviz.org%2Fworkflows%2Fexternal_request%2FUS_TN%2Finsert_contact_note%22%0Atimestamp%3D%222022-12-06T23:49:22.026643Z%22%0AinsertId%3D%22fjguw8g28xi7w9%22;timeRange=P7D?project=recidiviz-staging).
The request URL was
https://app-staging.recidiviz.org/workflows/external_request/US_TN/insert_contact_note

The easiest thing for us to do is disable this particular firewall rule,
which does the following: `Detects MySQL comment-/space-obfuscated
injections and backtick termination`. Each firewall rules is rated
against 4 levels of paranoia, with each higher rule being increasingly
sensitive with a potential for catching more malicious traffic but also
generating more false positives. This particular rule is level 2

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/17133

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f492de4bb00097f7e39e09fbbb9828242bff42e2