[US_ME] Added violations for Almost Eligible - SCCP Workflows (Recidiviz/recidiviz-data#17162)
## Description of the change

People that are only ineligible because of having a class A or B
violation in the previous 90 days will be surfaced in the form_query.

This currently contains 48 additional surfaced people.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 92d5e3d647ae51a668f231e5ef77bce37b31557f