[Ingest] Make ingest instance timestamps timezone aware (Recidiviz/recidiviz-data#17079)
## Description of the change

Carole and I were both seeing issues with running ingest instance unit
tests that dealt with timestamps locally. This PR introduces a VERY
simple test that writes a status then reads it again and asserts that
the timestamp we read is the timestamp we expect. This test was failing
locally but not in CI. This could be related to platform-specific quirks
of how Macs / SQLAlchemy interact with Postgres, but we are not sure.
The issues persist when we use Postgres 13 locally.

It appears that local bugginess goes away when you use the timezone
aware timestamp type in the postgres schema itself. This PR converts to
using a timezone aware timestamp, along with an accompanying migration.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Will allow us to simplify logic in Recidiviz/recidiviz-data#16986

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3091a2cd49459f2035acf56532cb55505df00229