Justice Counts] [2/n] Update user agencies in db when fetching user permissions (Recidiviz/recidiviz-data#17179)
## Description of the change

This PR adds a change where whenever the frontend sends a `PUT` request
to the `/users` endpoint, the BE will update the agencies associated
with the user. The frontend makes this request during every session in
order to fetch the right permissions for a user.

## Related issues

Related to Recidiviz/recidiviz-data#17156

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d759e45012d6831097857abd034bbb9779564add