Bug fix in TN sentence processing (Recidiviz/recidiviz-data#17033)
## Description of the change
This PR fixes an issue that @DSharm identified where we were dropping
sentences in TN at some point in the process where sentences in
`sentences_preprocessed` are aggregated to sentence groups in
`sentence_imposed_group_summary`.

This issue stems from the occurrence of two things:
1. In TN we can have a SUPERVISION sentence and an INCARCERATION
sentence with the same `external_id` value
2. Sentences can be be labeled as being consecutive to sentences of a
different type, but we do not consider that when constructing imposed
groups

**Example**
As an example, let's say a person is shown as having the following
sentences from `state_supervision_sentence` and
`state_incarceration_sentence`.

| Internal ID | External ID | Imposed Date | Sentence Type | Consecutive
To |
| -------------------- | -------------------- |--------------------
|-------------------- |-------------------- |
|	1	|	A	|	2020-01-01	|	Incarceration	|		|
|	2	|	A	|	2020-01-01	|	Supervision	|		|
|	3	|	B	|	2020-06-01	|	Supervision	|	A	|

In `sentences_preprocessed`, sentence 2 will never appear because we
deduplicate based on external id, prioritizing the incarceration
sentence. This itself isn't the issue that was identified but does
related to it.

In `us_tn_consecutive_sentences_preprocessed`, the output would
currently be the following.
|	Internal ID	|	Sentence Type	|	Consecutive To	|
|	--------------------	|	--------------------	|--------------------	|
|	1	|	Incarceration	|		|
|	2	|	Supervision	|		|
|	3	|	Supervision	|	2	|

In the above, Sentence 3 is being listed as consecutive to Sentence 2
and not Sentence 1 because Sentence 2 is of the same sentence type
(supervision). It is assumed (when imposed groups are constructed), that
every sentence is either a parent sentence (has a null value in
`consecutive_sentence_id`) or is a child sentence (has a value in
`consecutive_sentence_id` that links back to another `sentence_id`.
However, because sentence 2 does not exist in `sentences_preprocessed`,
sentence 3 is being dropped because it has a parent sentence that
doesn't exist. This means that the output in
`sentence_imposed_group_summary` would currently only include sentence
1, which is wrong.

Therefore, the change that I implemented was to deduplicate
sentences`us_tn_consecutive_sentences_preprocessed` based on
`external_id` (prior to calculating the consecutive id field), using the
same logic that is implemented in `us_tn_sentences_preprocessed`
(prioritizing the incarceration sentence).

That would result in the following output in
`us_tn_consecutive_sentences_preprocessed` where sentence 3 is no longer
listed as being consecutive to any other sentence. This is consistent
with how it is currently handled in other states if a supervision
sentence is listed as being consecutive to an incarceration sentence (we
don't consider it for total sentence length calculations).
|	Internal ID	|	Sentence Type	|	Consecutive To	|
|	--------------------	|	--------------------	|--------------------	|
|	1	|	Incarceration	|		|
|	3	|	Supervision	|		|

And then the resulting output in `sentence_imposed_group_summary` would
then include two groups (1) Sentence 1 imposed on 2020-01-01 and (2)
Sentence 3 imposed on 2020-06-01.

**Testing Output**
Output is in the sandbox `agaidus_tn_`. I've checked the examples in
[this](https://docs.google.com/spreadsheets/d/1X7qk8-A1xLsBRAjq8pZjE9qDWoytWyt9zZUrIfuio6c/edit?usp=sharing)
sheet and have included the `sentence_span` output from my sandbox.
@DSharm could you confirm that those outputs are what we would expect?

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#16991

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 097863d6725391156f974b37c2870d7f0c85afac