[Security] Lowering WAF's sensitivity to SQL Injection attacks (Recidiviz/recidiviz-data#17220)
## Description of the change

> The WAF continues to be overly sensitive to request headers when it
comes to potential SQL Injection attacks. GCP supports ignoring certain
request headers, but that is not currently supported in the main GCP
Terraform provider. It is only supported in beta at this time.

Given that limitation, I am going to lower the sensitivity to see if
that eliminates some of the noise in the WAF around SQL injection.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/17133

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 025d8efc52d8a26ae98240c3911f31e8d79865ad