[US_ME] Removed completed terms and flagged pending violations (Recidiviz/recidiviz-data#17232)
## Description of the change

> - When deciding between sub-sessions, prioritize active and completed
terms to get the right `end_date.`
> - Flagged pending disciplines with their respective time of creation
to know when they've been pending for a while.

The number of clients eligible dropped to 202 from 228.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7fc5b9d81889fc67d427c0b2ce00c181bac6c392