[Justice Counts] Script to backfill user agencies (Recidiviz/recidiviz-data#17227)
## Description of the change

This script will be run to backfill the `agencies` relationship in the
`UserAccount` table.

I'm not sure if this needs to be merged / a permanent part of this repo
since we will be doing this only once. I plan to follow up with
@lilidworkin during our 1:1 to get some guidance on how to run this in
the staging / production environments.

## Related issues

Closes Recidiviz/recidiviz-data#17208

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: aa9b3d7adb120aac239dc990068a2ced15355799